In driver/feature_scheduler_target.py, cast exposure times as float when initializing the parent class.
The FBS observation is a wrapper around numpy array and exposure times values are numpy floats. This ensures the values are python floats instead.

+++ python/lsst/ts/scheduler/driver/feature_scheduler_target.py
@@ -57,7 +57,7 @@ class FeatureSchedulerTarget(DriverTarget):
             ang_rad=observation["rotSkyPos"][0],
             num_exp=observation["nexp"][0],
             exp_times=[
-                observation["exptime"][0] / observation["nexp"][0]
+                float(observation["exptime"][0] / observation["nexp"][0])
                 for i in range(observation["nexp"][0])
             ],
         )
